+++ dbtransfers/get_variables.php
@@ -16,11 +16,8 @@ $statement->execute(array(':id' => $id));
 while($rs = $statement->fetch())//"rs" = "row statements"
 {
     $username = $rs['username'];//the variable username will contain the username of the last item of the array parsed
-
     $energy = $rs['energy'];
-
     $money = $rs['money'];
-
     $nbreplays = $rs['nbreplays'];
 }
 
@@ -35,7 +32,6 @@ $statement->execute(array(':id' => $id));
 while($rs = $statement->fetch())//"rs" = "row statements"
 {
     $story_location = $rs['storylocation'];//the variable username will contain the username of the last item of the array parsed
-
     $last_chapter = $rs['lastchapterplayed'];
 }
 
@@ -50,86 +46,53 @@ $statement->execute(array(':id' => $id));
 while($rs = $statement->fetch())//"rs" = "row statements"
 {
     $a1 = $rs['karolina'];//the variable username will contain the username of the last item of the array parsed
-
     $a2 = $rs['ellie'];
-
     $a3 = $rs['neha'];
-
     $a4 = $rs['raquel'];
-
     $a5 = $rs['claire'];
-
     $a6 = $rs['alistair'];
-
     $a7 = $rs['tadashi'];
-
     $a8 = $rs['tegan'];
-
     $a9 = $rs['tyler'];
-
     $a10 = $rs['axel'];
-
     $a11 = $rs['ladyarlington'];
-
     $a12 = $rs['coachdavis'];
-
     $a13 = $rs['serena'];
-
     $a14 = $rs['cecile'];
-
     $a15 = $rs['teacherchapter2'];
 }
 
 ?>
 
+<!-- NOW WE NEED TO OUTPUT ALL OF THIS TO HTML TO PASS IT TO THE JS CODE -->
+
 <html>
 
-    <!-- ***** -->
-
+    <!-- USERINFO table -->
     <div id="db_handle_username"><?php echo $username; ?></div>
-
     <div id="db_handle_energy"><?php echo $energy; ?></div>
-
     <div id="db_handle_money"><?php echo $money; ?></div>
-
     <div id="db_handle_nbreplays"><?php echo $nbreplays; ?></div>
 
-    <!-- ***** -->
-
+    <!-- STORY table -->
     <div id="db_handle_story_location"><?php echo $story_location; ?></div>
-
     <div id="db_handle_last_chapter_played"><?php echo $last_chapter; ?></div>
 
-    <!-- ***** -->
-
-    <div id="db_handle_a1"><?php echo $a1; ?></div>
-
-    <div id="db_handle_a2"><?php echo $a2; ?></div>
-
-    <div id="db_handle_a3"><?php echo $a3; ?></div>
-
-    <div id="db_handle_a4"><?php echo $a4; ?></div>
-
-    <div id="db_handle_a5"><?php echo $a5; ?></div>
-
-    <div id="db_handle_a6"><?php echo $a6; ?></div>
-
-    <div id="db_handle_a7"><?php echo $a7; ?></div>
-
-    <div id="db_handle_a8"><?php echo $a8; ?></div>
-
-    <div id="db_handle_a9"><?php echo $a9; ?></div>
-
-    <div id="db_handle_a10"><?php echo $a10; ?></div>
-
-    <div id="db_handle_a11"><?php echo $a11; ?></div>
-
-    <div id="db_handle_a12"><?php echo $a12; ?></div>
-
-    <div id="db_handle_a13"><?php echo $a13; ?></div>
-
-    <div id="db_handle_a14"><?php echo $a14; ?></div>
-
-    <div id="db_handle_a15"><?php echo $a15; ?></div>
+    <!-- AFFINITY table -->
+    <input id="db_handle_a1" value=<?php echo $a1; ?>></input>
+    <input id="db_handle_a2" value= <?php echo $a2; ?>></input>
+    <input id="db_handle_a3" value = <?php echo $a3; ?>></input>
+    <input id="db_handle_a4" value = <?php echo $a4; ?>></input>
+    <input id="db_handle_a5" value = <?php echo $a5; ?>></input>
+    <input id="db_handle_a6" value = <?php echo $a6; ?>></input>
+    <input id="db_handle_a7" value = <?php echo $a7; ?>></input>
+    <input id="db_handle_a8" value = <?php echo $a8; ?>></input>
+    <input id="db_handle_a9" value = <?php echo $a9; ?>></input>
+    <input id="db_handle_a10" value = <?php echo $a10; ?>></input>
+    <input id="db_handle_a11" value = <?php echo $a11; ?>></input>
+    <input id="db_handle_a12" value = <?php echo $a12; ?>></input>
+    <input id="db_handle_a13" value = <?php echo $a13; ?>></input>
+    <input id="db_handle_a14" value = <?php echo $a14; ?>></input>
+    <input id="db_handle_a15" value = <?php echo $a15; ?>></input>
 
 </html>