+++ profile/characterGenerator.php
@@ -14,6 +14,24 @@ function pullSex()
 	
 	// echo "storylocation = " . $storylocation . "; lastchapterplayed = " . $lastchapterplayed;
 		
+	//SEX
+	try
+	{
+		$skincolor = $_POST['sex'];
+
+		// $sqlQuery1 = "UPDATE scholarinfo SET scholar_sex = '$sex' WHERE id = '$id'";
+		// $statement = $db->prepare($sqlQuery1);
+		// $statement->execute();
+
+		// $status = "SQL Sent!";
+	}
+	catch (PDOException $ex)
+	{
+		// echo $ex;
+	}
+	
+	
+	/*
 	if (isset($_POST["sex"]))
 	{
 		$sex = $_POST['sex'];//WHY DOESN'T IT WORK
@@ -26,7 +44,9 @@ function pullSex()
 	else
 	{
 		return 0;
-	}
+	}*/
+
+
 
 	//if (isset($_POST ["address"])) if ($_POST ["address"] != "") { echo "Your form submission has an error."; exit; }	
 	
@@ -37,31 +57,62 @@ function pullSex()
 
 function generateImg()
 {
-	//Generate Random Numbers
-	$sex = pullSex(); //Generates a random number between 0 and 1; 0 being female and 1 being male
-	$skin = rand(0,4); //Generates a random number between 0 and 4; all corresponding to one of the choices of skin color
-	$hairstyle = rand(0,3); //same logic applies to the rest
-	$haircolor = rand(0,3);
-	$eyecolor = rand(0,3);
-	
-	//Generate a random body
-	$skin_rgb = getSkinColor($skin);
+	session_start();
+
+	if(isset($_SESSION['id']))
+	{
+		$id = $_SESSION['id']; // Picks up the id of the user that is logged in
+	}
+
+	try
+	{
+		$db = new PDO('mysql:host=localhost; dbname=USERS', 'root', 'root');
+		$db->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+		//echo "Connected to the database!";
+	}
+	catch(PDOException $ex)
+	{
+		//display error message
+		// echo "Connection failed ".$ex->getMessage();
+	}
+
+	$sqlQuery = "SELECT * FROM scholarinfo WHERE id = :id";
+	$statement = $db->prepare($sqlQuery);
+	$statement->execute(array(':id' => $id));
+
+	while($rs = $statement->fetch())
+	{
+		$sex = $rs['scholar_sex'];
+		$skincolor = $rs['scholar_skincolor'];
+		$hairstyle = $rs['scholar_hairstyle'];
+		$haircolor = $rs['scholar_haircolor'];
+		$eyecolor = $rs['scholar_eyecolor'];
+		$shirt = $rs['shirt_id'];
+		$pants = $rs['pants_id'];
+		$shoes = $rs['shoes_id'];
+	}
+
+	$skin_rgb = getSkinColor($skincolor);
 	$hair_rgb = getHairColor($haircolor);
 	$eye_rgb = getEyeColor($eyecolor);
 	$hair_number = getHairStyle($sex, $hairstyle, false);
 	$genericVariables = getGenericVariables($sex, false);
-	
-	/*****DEBUG*****
-		echo "Skin: R: " . $skin_rgb[0] . "; G: " . $skin_rgb[1] . "; B: " . $skin_rgb[2] . "; OP: " . $skin_rgb[3];
-		echo nl2br("  \n");
-		echo "Hair: R: " . $hair_rgb[0] . "; G: " . $hair_rgb[1] . "; B: " . $hair_rgb[2] . "; OP: " . $hair_rgb[3];
-		echo nl2br("  \n");
-		echo "Eyes: R: " . $eye_rgb[0] . "; G: " . $eye_rgb[1] . "; B: " . $eye_rgb[2] . "; OP: " . $eye_rgb[3];
-		echo nl2br("  \n");
-		echo "Hairstyle: " . $hair_number;
-	*/
-	
-	//Generate the image
+	// GET PANTS
+	// GET SHOES
+	// etc. (must be simplified to the lowest level)
+
+	//SEX
+    try
+    {
+        // $skincolor = $_POST['sex'];//this line doesn't work
+    }
+    catch (Exception $ex)
+    {
+        echo $ex;
+    }
+
+
+
 	header('Content-Type: image/png');
 
 	$destination = imagecreatefrompng($genericVariables[0]);//body type
@@ -70,6 +121,7 @@ function generateImg()
 	$source3 = imagecreatefrompng($genericVariables[3]);//lips image
 	$source4 = imagecreatefrompng($hair_number);//hair style
 	$source5 = imagecreatefrompng($genericVariables[4]);//uniform
+	$source6 = imagecreatefrompng($genericVariables[5]);//shoes
 
 	imagealphablending($source1, true);
 	imagesavealpha($source1, true);
@@ -77,8 +129,8 @@ function generateImg()
 	imagealphablending($destination, true);
 	imagesavealpha($destination, true);
 
-	$largeur_source1 = imagesx($source1);
-	$hauteur_source1 = imagesy($source1);
+	$largeur_source = imagesx($source1);
+	$hauteur_source = imagesy($source1);
 
 	$largeur_source2 = imagesx($source2);
 	$hauteur_source2 = imagesy($source2);
@@ -92,19 +144,23 @@ function generateImg()
 	$largeur_source5 = imagesx($source5);
 	$hauteur_source5 = imagesy($source5);
 
-	//Apply the colors
+	$largeur_source6 = imagesx($source6);
+	$hauteur_source6 = imagesy($source6);
+
 	imagefilter($source2, IMG_FILTER_COLORIZE, $eye_rgb[0], $eye_rgb[1], $eye_rgb[2], $eye_rgb[3]); //Eye color
 	imagefilter($source3, IMG_FILTER_COLORIZE, $skin_rgb[0], $skin_rgb[1], $skin_rgb[2], $skin_rgb[3]); //Lip color = skin color
 	imagefilter($source4, IMG_FILTER_COLORIZE, $hair_rgb[0], $hair_rgb[1], $hair_rgb[2], $hair_rgb[3]); //Hair color
 	imagefilter($destination, IMG_FILTER_COLORIZE, $skin_rgb[0], $skin_rgb[1], $skin_rgb[2], $skin_rgb[3]);//Skin color
 
-	imagecopy($destination, $source1, 0, 0, 0, 0, $largeur_source1, $hauteur_source1);
+
+	imagecopy($destination, $source1, 0, 0, 0, 0, $largeur_source, $hauteur_source);
 	imagecopy($destination, $source2, 0, 0, 0, 0, $largeur_source2, $hauteur_source2);
 	imagecopy($destination, $source3, 0, 0, 0, 0, $largeur_source3, $hauteur_source3);
 	imagecopy($destination, $source4, 0, 0, 0, 0, $largeur_source4, $hauteur_source4);
 	imagecopy($destination, $source5, 0, 0, 0, 0, $largeur_source5, $hauteur_source5);
+	imagecopy($destination, $source6, 0, 0, 0, 0, $largeur_source6, $hauteur_source6);
 
-	imagepng($destination);
+	imagepng($destination);	
 }
 
 generateImg();//void function to generate the img, the entire php file is going to become an img (*.png) file